+++ src/main.jsx
@@ -3,9 +3,10 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import './index.css'
 import App from './App'
+import HostInfo from './pages/host_info/HostInfo'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
+    <HostInfo />
   </React.StrictMode>,
 )
