+++ src/pages/host_info/Item2Component.jsx
